# swagger.yaml
openapi: 3.0.0
info:
    title: User API
    version: 1.0.0
    description: API Documentation for retrieving users
servers:
    - url: http://localhost:3000
paths:
...
    /users/{id}:
        get:
            summary: Get a user by ID
            description: Retrieve a specific user by their unique ID
            parameters:
                - name: id
                  in: path
                  required: true
                  description: Numeric ID of the user to retrieve
                  schema:
                      type: integer
                      example: 2
            responses:
                200:
                    description: The requested user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                404:
                    description: User not found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: "User not found"
...
components:
    schemas:
        BookUserData:
            type: object
            properties:
                id:
                    type: string
                    example: "book123"
                reason:
                    type: string
                    example: "Loved the storyline"
        Message:
            type: object
            properties:
                id:
                    type: integer
                    example: 1
                type:
                    type: string
                    example: "friend_request"
        User:
            type: object
            properties:
                username:
                    type: string
                    example: "john_doe"
                id:
                    type: integer
                    example: 123
                likes:
                    type: array
                    items:
                        $ref: "#/components/schemas/BookUserData"
                dislikes:
                    type: array
                    items:
                        $ref: "#/components/schemas/BookUserData"
                pending:
                    type: array
                    items:
                        type: integer
                    example: [101, 102]
                friends:
                    type: array
                    items:
                        type: integer
                    example: [201, 202]
                inbox:
                    type: array
                    items:
                        $ref: "#/components/schemas/Message"
...